+++ NEWS.md
@@ -2,10 +2,12 @@
 
 #### New features
 
+* Extended [`add_constraint_boundary()`] with a new `method = "zone"` option for per-timestep spatial zone masking during scenario projections.
+* Replacement of the `st_kde` function with an internal implementation of kernel density estimation in c++ for speed.
+
 #### Minor improvements and bug fixes
 
 * Minor :bug: fixes with scenario predictons not accounting properly for zone limits.
-* Replacement of the `st_kde` function with an internal implementation of kernel density estimation in c++
 * Fixed Log function for both model and scenario fitting.
 
 # ibis.iSDM 0.1.7